+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 *CURRENT*
 ---
     
+- 2017-12-15 (Laisky) feat: add ''DFAFilter'' -> [view commit](https://github.com/Laisky/kipp/commit//9e386a289d0c5ed0f3c0333b792502aa36c45e61)
+- 2017-11-07 (Laisky) ci: changelog -> [view commit](https://github.com/Laisky/kipp/commit//436e722efdc0fd6075ae918122fd38c040fe22b2)
 - 2017-11-07 (Laisky) fix: remove some movoto functions -> [view commit](https://github.com/Laisky/kipp/commit//ea409470b7f2ca7e67bb64a539b2f2379751da1f)
 - 2017-11-07 (Laisky) fix: EmailSender support username and passwd -> [view commit](https://github.com/Laisky/kipp/commit//dd6ddd842cd3fe608b7aecfb1eb6987ba6b9a331)
 - 2017-11-07 (Laisky) ci: changelog -> [view commit](https://github.com/Laisky/kipp/commit//b1d726eabb088133b3b9dfe3c9c1e05db94c8a51)